Install npm module mime
Using command `npm install mime --save` to add mime type module
Get the mime type associated with a file, if no mime type is found
application/octet-stream is returned

MIME stands for Multipurpose Internet Mail Extensions.
It's a way of identifying files on the Internet according
to their nature and format.

For example, using the Content-type header value defined in a HTTP
response, the browser can open the file with the proper
extension/plugin.