Builds out full filepath, folder, and extension
Script now uploads a file to a folder based on today's date.
The filepath is getting calculated based on mime.lookup (installed
as a dependency) to include the file extension.

ContentType is being sent to AWS so that our AWS
knows how to handle the link to our img/file.